Crate namada_io

Traits for implementing IO handlers. This is to enable generic IO. The defaults are the obvious Rust native functions.

Re-exports

pub use client::Client;

Modules

client

Macros

display

Convenience macro for formatting arguments to Io::print

display_line

Convenience macro for formatting arguments to Io::println and Io::writeln

edisplay_line

Convenience macro for formatting arguments to Io::eprintln

prompt

A convenience macro for formatting the user prompt before forwarding it to the Io::prompt method.

Structs

DevNullProgressBar

NOOP progress bar implementation.

NullIo

Ignores all I/O operations.

StdIo

Rust native I/O handling.

Traits

Io

A trait that abstracts out I/O operations

MaybeSend

MaybeSync

NamadaIo

ProgressBar

Mechanism that allows keeping track of the progress of some operation.

Functions

prompt_aux

A generic function for displaying a prompt to users and reading in their response.

read_aux

A generic function for reading input from users