namada_io::client
Trait Client

pub trait Client {
    type Error: From<Error> + Display + Debug;

    // Required methods
    fn request<'life0, 'async_trait>(
        &'life0 self,
        path: String,
        data: Option<Vec<u8>>,
        height: Option<BlockHeight>,
        prove: bool,
    ) -> Pin<Box<dyn Future<Output = Result<EncodedResponseQuery, Self::Error>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn perform<'life0, 'async_trait, R>(
        &'life0 self,
        request: R,
    ) -> Pin<Box<dyn Future<Output = Result<R::Output, RpcError>> + 'async_trait>>
       where R: SimpleRequest + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn simple_request<'life0, 'async_trait>(
        &'life0 self,
        path: String,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, Self::Error>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn abci_info<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Info, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn broadcast_tx_sync<'life0, 'async_trait>(
        &'life0 self,
        tx: impl 'async_trait + Into<Vec<u8>> + MaybeSend,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn latest_block<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn block<'life0, 'async_trait, H>(
        &'life0 self,
        height: H,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where H: TryInto<Height> + Send + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn block_search<'life0, 'async_trait>(
        &'life0 self,
        query: Query,
        page: u32,
        per_page: u8,
        order: Order,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn block_results<'life0, 'async_trait, H>(
        &'life0 self,
        height: H,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where H: TryInto<Height> + Send + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn tx_search<'life0, 'async_trait>(
        &'life0 self,
        query: Query,
        prove: bool,
        page: u32,
        per_page: u8,
        order: Order,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn abci_query<'life0, 'async_trait, V>(
        &'life0 self,
        path: Option<String>,
        data: V,
        height: Option<Height>,
        prove: bool,
    ) -> Pin<Box<dyn Future<Output = Result<AbciQuery, RpcError>> + 'async_trait>>
       where V: Into<Vec<u8>> + Send + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn latest_block_results<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn blockchain<'life0, 'async_trait, H>(
        &'life0 self,
        min: H,
        max: H,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where H: TryInto<Height> + Send + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn commit<'life0, 'async_trait, H>(
        &'life0 self,
        height: H,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where H: TryInto<Height> + Send + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn consensus_params<'life0, 'async_trait, H>(
        &'life0 self,
        height: H,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where H: TryInto<Height> + Send + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn consensus_state<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn latest_consensus_params<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn latest_commit<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn health<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<(), RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn net_info<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn status<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }

}

A client with async request dispatcher method, which can be used to invoke type-safe methods from a root queries Router, generated via router! macro.

Required Associated Types

type Error: From<Error> + Display + Debug

std::io::Error can happen in decoding with BorshDeserialize::try_from_slice

Required Methods

fn request<'life0, 'async_trait>( &'life0 self, path: String, data: Option<Vec<u8>>, height: Option<BlockHeight>, prove: bool, ) -> Pin<Box<dyn Future<Output = Result<EncodedResponseQuery, Self::Error>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Send a query request at the given path.

fn perform<'life0, 'async_trait, R>( &'life0 self, request: R, ) -> Pin<Box<dyn Future<Output = Result<R::Output, RpcError>> + 'async_trait>>
where R: SimpleRequest + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

Perform a request against the RPC endpoint

Provided Methods

fn simple_request<'life0, 'async_trait>( &'life0 self, path: String, ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, Self::Error>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Send a simple query request at the given path. For more options, use the request method.

fn abci_info<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Info, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/abci_info: get information about the ABCI application.

fn broadcast_tx_sync<'life0, 'async_trait>( &'life0 self, tx: impl 'async_trait + Into<Vec<u8>> + MaybeSend, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/broadcast_tx_sync: broadcast a transaction, returning the response from CheckTx.

fn latest_block<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/block: get the latest block.

fn block<'life0, 'async_trait, H>( &'life0 self, height: H, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where H: TryInto<Height> + Send + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

/block: get block at a given height.

fn block_search<'life0, 'async_trait>( &'life0 self, query: Query, page: u32, per_page: u8, order: Order, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/block_search: search for blocks by BeginBlock and EndBlock events.

fn block_results<'life0, 'async_trait, H>( &'life0 self, height: H, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where H: TryInto<Height> + Send + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

/block_results: get ABCI results for a block at a particular height.

fn tx_search<'life0, 'async_trait>( &'life0 self, query: Query, prove: bool, page: u32, per_page: u8, order: Order, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/tx_search: search for transactions with their results.

fn abci_query<'life0, 'async_trait, V>( &'life0 self, path: Option<String>, data: V, height: Option<Height>, prove: bool, ) -> Pin<Box<dyn Future<Output = Result<AbciQuery, RpcError>> + 'async_trait>>
where V: Into<Vec<u8>> + Send + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

/abci_query: query the ABCI application

fn latest_block_results<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/block_results: get ABCI results for the latest block.

fn blockchain<'life0, 'async_trait, H>( &'life0 self, min: H, max: H, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where H: TryInto<Height> + Send + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

/blockchain: get block headers for min <= height <= max.

Block headers are returned in descending order (highest first).

Returns at most 20 items.

fn commit<'life0, 'async_trait, H>( &'life0 self, height: H, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where H: TryInto<Height> + Send + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

/commit: get block commit at a given height.

fn consensus_params<'life0, 'async_trait, H>( &'life0 self, height: H, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where H: TryInto<Height> + Send + 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

/consensus_params: get current consensus parameters at the specified height.

fn consensus_state<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/consensus_state: get current consensus state

fn latest_consensus_params<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/consensus_params: get the latest consensus parameters.

fn latest_commit<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/commit: get the latest block commit

fn health<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<(), RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/health: get node health.

Returns empty result (200 OK) on success, no response in case of an error.

fn net_info<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/net_info: obtain information about P2P and other network connections.

fn status<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Response, RpcError>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/status: get Tendermint status including node info, pubkey, latest block hash, app hash, block height and time.

Object Safety

This trait is not object safe.