namada_io
Trait Io

pub trait Io {
    // Provided methods
    fn print(&self, output: impl AsRef<str>) { ... }
    fn flush(&self) { ... }
    fn println(&self, output: impl AsRef<str>) { ... }
    fn write<W: Write>(&self, writer: W, output: impl AsRef<str>) -> Result<()> { ... }
    fn writeln<W: Write>(
        &self,
        writer: W,
        output: impl AsRef<str>,
    ) -> Result<()> { ... }
    fn eprintln(&self, output: impl AsRef<str>) { ... }
    fn read<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<String>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn prompt<'life0, 'async_trait>(
        &'life0 self,
        question: impl 'async_trait + AsRef<str> + MaybeSync + MaybeSend,
    ) -> Pin<Box<dyn Future<Output = String> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
}

A trait that abstracts out I/O operations

Provided Methods

fn print(&self, output: impl AsRef<str>)

Print the given string

fn flush(&self)

Flush the output

fn println(&self, output: impl AsRef<str>)

Print the given string with a newline

fn write<W: Write>(&self, writer: W, output: impl AsRef<str>) -> Result<()>

Print the given string into the given Writer

fn writeln<W: Write>(&self, writer: W, output: impl AsRef<str>) -> Result<()>

Print the given string into the given Writer and terminate with newline

fn eprintln(&self, output: impl AsRef<str>)

Print the given error string

fn read<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<String>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Read a string from input

fn prompt<'life0, 'async_trait>( &'life0 self, question: impl 'async_trait + AsRef<str> + MaybeSync + MaybeSend, ) -> Pin<Box<dyn Future<Output = String> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Display the given prompt and return the string input

Object Safety

This trait is not object safe.