namada_io
Trait ProgressBar

pub trait ProgressBar {
    // Required methods
    fn upper_limit(&self) -> u64;
    fn set_upper_limit(&mut self, limit: u64);
    fn increment_by(&mut self, amount: u64);
    fn message(&mut self, message: String);
    fn finish(&mut self);
}

Mechanism that allows keeping track of the progress of some operation.

Required Methods

fn upper_limit(&self) -> u64

Query the amount of virtual elements the progress tracker is tracking.

fn set_upper_limit(&mut self, limit: u64)

Set the amount of virtual elements the progress tracker is tracking.

fn increment_by(&mut self, amount: u64)

Announce that amount virtual elements have been processed.

fn message(&mut self, message: String)

Specify a message to be display alongside the bar.

fn finish(&mut self)

Signal that a progress bar has completed